Packages Update

+++ waybar/script/launch.sh
@@ -1,3 +1,9 @@
-pkill waybar
-sleep 0.5
-waybar
+#!/bin/bash
+
+if pgrep -x "waybar" > /dev/null; then
+  # If Waybar is running, kill it
+  pkill waybar
+else
+  # If Waybar is not running, start it
+  waybar &
+fi
